Fixed bug with vertical lines

+++ Data Structures and Algorithms/src/algorithms_Course_1_MergeSort_QuickSort/Point.java
@@ -42,7 +42,7 @@ public class Point implements Comparable<Point>{
         if(that.y==this.y && that.x==this.x) return Double.NEGATIVE_INFINITY;
     	else if(that.y==this.y) return 0.0;
         else if(that.x==this.x) return Double.POSITIVE_INFINITY;
-		return ((double)(that.y-this.y))/((that.x-this.x));
+		return ((double)(that.y-this.y))/((double)(that.x-this.x));
     }
 
     // is this point lexicographically smaller than that one?
